move `givens_decomposition` and `_reduced_row_echelon` from `qchem` to `math` module (#7545)
**Context:**

These are purely mathematical helper functions and should therefore
belong to the `math` module.

**Description of the Change:**

- Move helpers to `math` module.
- Renamed RRE and made it visible in documentation

**Benefits:** Will help remove some weird dependencies on `qchem`.

**Possible Drawbacks:** None identified.

[sc-92361]

---------
