+++ doc/introduction/chemistry.rst
@@ -265,7 +265,7 @@ Utility functions
     ~pennylane.qchem.excitations
     ~pennylane.qchem.excitations_to_wires
     ~pennylane.qchem.factorize
-    ~pennylane.qchem.givens_decomposition
+    ~pennylane.math.givens_decomposition
     ~pennylane.qchem.hf_state
     ~pennylane.import_operator
     ~pennylane.qchem.import_state
