+++ pennylane/qchem/__init__.py
@@ -14,13 +14,13 @@
 """
 This submodule provides the functionality to perform quantum chemistry calculations.
 """
+from pennylane.math.decomposition import givens_decomposition
 from .basis_data import load_basisset
 from .basis_set import BasisFunction, atom_basis_data, mol_basis_data
 from .convert import import_operator, import_state
 from .convert_openfermion import from_openfermion, to_openfermion
 from .dipole import dipole_integrals, dipole_moment, fermionic_dipole, molecular_dipole
 from .factorization import basis_rotation, factorize, symmetry_shift
-from .givens_decomposition import givens_decomposition
 from .hamiltonian import (
     electron_integrals,
     fermionic_hamiltonian,
